Worked on refactoring Solitaire, created reset statistics button and all scripting necessary for it, and created written test cases for news story and reset statistics story
Refactoring is still not done, I am stuck on a certain issue which is causing many problems. Time for this commit might seem high but its all the time I have spent this week working on refactoring without committing any changes because no changes made any progress. On top of that, the time it took to create the reset statistics button and all scripting, plus the time it took to create the written test cases for 2 stories.